The cacheAge of the OembedResponse has been changed to long for compatibility with twitters large cache age values.

+++ src/main/java/ac/simons/oembed/OembedBuilder.java
@@ -74,7 +74,7 @@ public class OembedBuilder {
 		return this;
 	}
 
-	public OembedBuilder withIgnoreFailedUrlsForSeconds(Integer ignoreFailedUrlsForSeconds) {
+	public OembedBuilder withIgnoreFailedUrlsForSeconds(Long ignoreFailedUrlsForSeconds) {
 		oembed.setIgnoreFailedUrlsForSeconds(ignoreFailedUrlsForSeconds);
 		return this;
 	}
